Move `FieldPaths` utility to a separate file
